// packages/datadog-api-client-v1/models/SyntheticsVariableParser.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SyntheticsGlobalVariableParserType } from './SyntheticsGlobalVariableParserType';
/**
* Details of the parser to use for the global variable.
*/
export declare class SyntheticsVariableParser {
    'type': SyntheticsGlobalVariableParserType;
    /**
    * Regex or JSON path used for the parser. Not used with type `raw`.
    */
    'value'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SyntheticsVariableParser;
    static serialize(data: SyntheticsVariableParser): {
        [key: string]: any;
    };
    constructor();
}